change the types of DB to align with app

+++ academy-enterprise.sql
@@ -21,8 +21,10 @@ $$ LANGUAGE plpgsql;
 -- =============================================
 
 -- TABLA BASE DE AUTENTICACIÓN
+-- DECISIÓN: Usar BIGSERIAL para compatibilidad con Long en Java
+-- RAZÓN: Evita desbordamiento (2B vs 9 quintillones) y consistencia con JPA
 CREATE TABLE usuario (
-    id_usuario SERIAL PRIMARY KEY,
+    id_usuario BIGSERIAL PRIMARY KEY,
     email VARCHAR(150) UNIQUE NOT NULL,
     password_hash VARCHAR(255) NOT NULL,
     rol VARCHAR(20) NOT NULL CHECK (rol IN ('ADMIN', 'PROFESOR', 'ALUMNO')),
@@ -34,7 +36,7 @@ CREATE TABLE usuario (
 
 -- TABLA DE DATOS PERSONALES (Relación 1:1, PII/GDPR)
 CREATE TABLE datos_personales (
-    id_usuario INT PRIMARY KEY REFERENCES usuario(id_usuario),
+    id_usuario BIGINT PRIMARY KEY REFERENCES usuario(id_usuario),
     nombre VARCHAR(100) NOT NULL,
     apellidos VARCHAR(100) NOT NULL,
     dni_nie VARCHAR(20) UNIQUE,
@@ -45,7 +47,7 @@ CREATE TABLE datos_personales (
     
     -- AUDITORÍA
     fecha_modificacion TIMESTAMP,
-    updated_by INT REFERENCES usuario(id_usuario)
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 
 -- TRIGGER para actualizar fecha_modificacion en datos_personales
@@ -67,7 +69,7 @@ WHERE u.rol = 'ALUMNO';
 
 -- TABLAS CON AUDITORÍA COMPLETA
 CREATE TABLE comunidad (
-    id_comunidad SERIAL PRIMARY KEY,
+    id_comunidad BIGSERIAL PRIMARY KEY,
     codigo VARCHAR(10) NOT NULL UNIQUE,
     nombre VARCHAR(100) NOT NULL,
     capital VARCHAR(100),
@@ -75,13 +77,13 @@ CREATE TABLE comunidad (
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_comunidad_update BEFORE UPDATE ON comunidad FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
 CREATE TABLE empresa (
-    id_empresa SERIAL PRIMARY KEY,
+    id_empresa BIGSERIAL PRIMARY KEY,
     cif VARCHAR(20) NOT NULL UNIQUE,
     nombre_legal VARCHAR(150) NOT NULL,
     direccion_fiscal TEXT,
@@ -89,85 +91,85 @@ CREATE TABLE empresa (
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_empresa_update BEFORE UPDATE ON empresa FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
 
 CREATE TABLE centro (
-    id_centro SERIAL PRIMARY KEY,
+    id_centro BIGSERIAL PRIMARY KEY,
     codigo_centro VARCHAR(20) UNIQUE NOT NULL,
     nombre VARCHAR(100) NOT NULL,
-    id_empresa INT REFERENCES empresa(id_empresa),
-    id_comunidad INT REFERENCES comunidad(id_comunidad),
+    id_empresa BIGINT REFERENCES empresa(id_empresa),
+    id_comunidad BIGINT REFERENCES comunidad(id_comunidad),
     capacidad_maxima INT CHECK (capacidad_maxima >= 0),
     activo BOOLEAN DEFAULT TRUE,
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_centro_update BEFORE UPDATE ON centro FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
 
 CREATE TABLE materia (
-    id_materia SERIAL PRIMARY KEY,
+    id_materia BIGSERIAL PRIMARY KEY,
     nombre VARCHAR(100) NOT NULL,
     activo BOOLEAN DEFAULT TRUE,
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_materia_update BEFORE UPDATE ON materia FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
 
 CREATE TABLE formato (
-    id_formato SERIAL PRIMARY KEY, 
+    id_formato BIGSERIAL PRIMARY KEY, 
     nombre VARCHAR(50) NOT NULL,
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_formato_update BEFORE UPDATE ON formato FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
 
 CREATE TABLE curso (
-    id_curso SERIAL PRIMARY KEY,
+    id_curso BIGSERIAL PRIMARY KEY,
     nombre VARCHAR(150) NOT NULL,
-    id_materia INT REFERENCES materia(id_materia),
-    id_formato INT REFERENCES formato(id_formato),
+    id_materia BIGINT REFERENCES materia(id_materia),
+    id_formato BIGINT REFERENCES formato(id_formato),
     precio_base DECIMAL(10, 2) NOT NULL CHECK (precio_base > 0), -- Constraint: precio > 0
     duracion_horas INT CHECK (duracion_horas > 0), -- Constraint: horas > 0
     activo BOOLEAN DEFAULT TRUE,
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_curso_update BEFORE UPDATE ON curso FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
 
 CREATE TABLE convocatoria (
-    id_convocatoria SERIAL PRIMARY KEY,
+    id_convocatoria BIGSERIAL PRIMARY KEY,
     codigo VARCHAR(50) UNIQUE,
-    id_curso INT REFERENCES curso(id_curso),
-    id_profesor INT REFERENCES usuario(id_usuario),
-    id_centro INT REFERENCES centro(id_centro),
+    id_curso BIGINT REFERENCES curso(id_curso),
+    id_profesor BIGINT REFERENCES usuario(id_usuario),
+    id_centro BIGINT REFERENCES centro(id_centro),
     fecha_inicio DATE NOT NULL,
     fecha_fin DATE NOT NULL CHECK (fecha_fin > fecha_inicio), -- Constraint: fin posterior a inicio
     activo BOOLEAN DEFAULT TRUE,
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_convocatoria_update BEFORE UPDATE ON convocatoria FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
@@ -177,23 +179,23 @@ CREATE TRIGGER trg_convocatoria_update BEFORE UPDATE ON convocatoria FOR EACH RO
 -- =============================================
 
 CREATE TABLE entidad_subvencionadora (
-    id_entidad SERIAL PRIMARY KEY,
+    id_entidad BIGSERIAL PRIMARY KEY,
     nombre VARCHAR(150),
     codigo_oficial VARCHAR(50),
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_entidad_update BEFORE UPDATE ON entidad_subvencionadora FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
 
 CREATE TABLE matricula (
-    id_matricula SERIAL PRIMARY KEY,
+    id_matricula BIGSERIAL PRIMARY KEY,
     codigo VARCHAR(50) UNIQUE,
-    id_convocatoria INT REFERENCES convocatoria(id_convocatoria),
-    id_alumno INT REFERENCES usuario(id_usuario),
+    id_convocatoria BIGINT REFERENCES convocatoria(id_convocatoria),
+    id_alumno BIGINT REFERENCES usuario(id_usuario),
     fecha_matricula TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     
     -- TRAZABILIDAD FINANCIERA
@@ -202,7 +204,7 @@ CREATE TABLE matricula (
     motivo_descuento VARCHAR(100),
     
     -- SUBVENCIONES
-    id_entidad_subvencionadora INT REFERENCES entidad_subvencionadora(id_entidad),
+    id_entidad_subvencionadora BIGINT REFERENCES entidad_subvencionadora(id_entidad),
     importe_subvencionado DECIMAL(10,2) DEFAULT 0 CHECK (importe_subvencionado >= 0),
     
     -- Campo calculado (PostgreSQL 12+)
@@ -213,8 +215,8 @@ CREATE TABLE matricula (
     -- AUDITORÍA
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 -- El trigger de precio se mantiene, pero ahora se combina con el created_by en el INSERT.
 CREATE TRIGGER trg_calcular_precio_matricula
@@ -227,9 +229,9 @@ CREATE TRIGGER trg_matricula_update BEFORE UPDATE ON matricula FOR EACH ROW EXEC
 
 -- Resto de tablas (Factura, Calificación) también llevan auditoría.
 CREATE TABLE factura (
-    id_factura SERIAL PRIMARY KEY,
+    id_factura BIGSERIAL PRIMARY KEY,
     numero_factura VARCHAR(50) UNIQUE NOT NULL,
-    id_matricula INT REFERENCES matricula(id_matricula),
+    id_matricula BIGINT REFERENCES matricula(id_matricula),
     fecha_emision TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     base_imponible DECIMAL(10,2) NOT NULL CHECK (base_imponible >= 0),
     iva_porcentaje DECIMAL(5,2) DEFAULT 21.00 CHECK (iva_porcentaje >= 0),
@@ -238,22 +240,22 @@ CREATE TABLE factura (
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_factura_update BEFORE UPDATE ON factura FOR EACH ROW EXECUTE FUNCTION update_timestamp();
 
 
 CREATE TABLE calificacion (
-    id_calificacion SERIAL PRIMARY KEY,
-    id_matricula INT REFERENCES matricula(id_matricula),
+    id_calificacion BIGSERIAL PRIMARY KEY,
+    id_matricula BIGINT REFERENCES matricula(id_matricula),
     nota DECIMAL(4, 2) CHECK (nota >= 0 AND nota <= 10), -- Constraint: nota entre 0 y 10
     comentarios TEXT,
     
     fecha_creacion TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion TIMESTAMP,
-    created_by INT REFERENCES usuario(id_usuario),
-    updated_by INT REFERENCES usuario(id_usuario)
+    created_by BIGINT REFERENCES usuario(id_usuario),
+    updated_by BIGINT REFERENCES usuario(id_usuario)
 );
 CREATE TRIGGER trg_calificacion_update BEFORE UPDATE ON calificacion FOR EACH ROW EXECUTE FUNCTION update_timestamp();
